+++ CHANGELOG.md
@@ -20,6 +20,9 @@ Release date: TBD
 - Add access to settings through tray menu
 - Removed unclear desktop notifications when failed to load tabs.
 - Reload automatically the failed tab when the computer becomes online.
+- Added back/forward features for the current tab.
+  - Windows and Linux: Alt+Left, Alt+Right
+  - OS X: Command+[, Command+]
 
 #### Windows
 - Added an option to toogle the red dot icon for unread messages (default is on).